Nueva clase Base.
