+++ depositos/deposito.py
@@ -1,9 +1,10 @@
 import json
 from pathlib import Path
+from depositos.base import Base
 from depositos.distrito import Distrito
 
 
-class Deposito(object):
+class Deposito(Base):
     """ Depósito """
 
     def __init__(self, config, ruta):
@@ -30,24 +31,6 @@ class Deposito(object):
                     distrito.rastrear()
                     self.distritos.append(distrito)
             self.ya_rastreado = True
-
-    def separar_fecha_descripcion(self, distrito, autoridad, archivo):
-        separados = archivo.name.split('-')
-        if len(separados) >= 3:
-            fecha = f'{separados[0]}-{separados[1]}-{separados[2]}'
-        else:
-            fecha = self.config.fecha_por_defecto
-        if len(separados) >= 4:
-            descripcion = ' '.join(separados[3:])
-        else:
-            descripcion = ''
-        return({'distrito': distrito.nombre, 'autoridad': autoridad.nombre, 'fecha': fecha, 'descripcion': descripcion, 'archivo': archivo.name})
-
-    def separar_fecha_expediente_descripcion(self, distrito, autoridad, archivo):
-        return({})
-
-    def separar_fecha_sentencia_expediente_genero_descripcion(self, distrito, autoridad, archivo):
-        return({})
 
     def crear_ruta_json(self):
         """ Crear la ruta al archivo JSON /var/www/html/consultas/v2.0/<RAMA>/<AAAA-MM-DD>.json """
@@ -78,15 +61,6 @@ class Deposito(object):
                 listado += [funcion(distrito, autoridad, archivo) for archivo in autoridad.archivos]
         salida = {'data': listado}
         return(json.dumps(salida))
-
-    def guardar_json(self):
-        ruta = self.crear_ruta_json()
-        padre_dir = ruta.parent
-        if not padre_dir.exists():
-            padre_dir.mkdir(parents=True)
-        with open(ruta, 'w') as puntero:
-            puntero.write(self.crear_contenido_json())
-        return(str(ruta))
 
     def __repr__(self):
         distritos_repr = '\n  '.join([repr(distrito) for distrito in self.distritos])
